service/requirements.txt
azure-servicebus==0.50.2
Flask==1.1.1
requests==2.22.0
sesamutils==0.1.10
